Use shell quotes around regex in apt search in doc
This PR, which is documentation-only and which does NOT include searching for other instances of this issue, addresses an issue I found when following the docs while using the `fish` shell. Without the quotes included in this PR, my shell reports `apt search ^regolith-looks-` returning no results with no errors. Including the quotes generates the expected result.

I don't know for sure but I believe the issue is that some shells treat `^` as a special character here as part of argument processing. It seems from my brief search that including the quotes doesn't break any other shells and should work in all cases.